modificationYasmine

+++ Phase-Spring/EstateMarketPi/src/main/java/pi/app/estatemarket/Services/ServiceComment.java
@@ -16,8 +16,10 @@ import pi.app.estatemarket.dto.CommentDTO;
 
 //import pi.app.estatemarket.dto.PublicationDTO;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -39,28 +41,36 @@ public class ServiceComment implements IServiceComment {
     }
 
     @Override
-    public Comment updateComment(int IdComment, Comment comm) {
+    public Comment updateComment(int IdComment, Comment comm) throws Exception{
         return commentRepository.findById(IdComment)
                 .map(comment1 -> {
                     comment1.setDescriptionCommentaire(comm.getDescriptionCommentaire());
                     return commentRepository.save(comment1);
-                }).orElseThrow(() -> new RuntimeException("Commentaire non trouvé !"));
+                }).orElseThrow(() -> new RuntimeException("Comment not found !"));
 }
     @Override
-    public Comment retrieveComment(Integer IdComment) {
-        return commentRepository.findById(IdComment).get();
+    public Comment retrieveComment(Integer IdComment) throws Exception {
+        Optional<Comment> commentOptional = commentRepository.findById(IdComment);
+        if (!commentOptional.isPresent()) {
+            throw new Exception("Comment with ID " + IdComment + " does not exist.");
+        }
+        return commentOptional.get();
     }
 
+
     @Override
-    public void removeComment(Integer IdComment) {
+    public void removeComment(Integer IdComment) throws Exception {
+        if (!commentRepository.existsById(IdComment)) {
+            throw new Exception("Comment with ID " + IdComment + " does not exist.");
+        }
         commentRepository.deleteById(IdComment);
     }
 
 
     @Override
-    public void ajouterEtAffecterCommentaireAUserEtCommentaire(Comment comment, Long userID, int IdPublication) {
-        UserApp user = userRepository.findById(userID).orElse(null);
-        Publication publication = publicationRepository.findById(IdPublication).orElse(null);
+    public void ajouterEtAffecterCommentaireAUserEtCommentaire(Comment comment, Long userID, int IdPublication) throws Exception {
+        UserApp user = userRepository.findById(userID).orElseThrow(() -> new Exception("User with ID " + userID + " does not exist."));
+        Publication publication = publicationRepository.findById(IdPublication).orElseThrow(() -> new Exception("Publication with ID " + IdPublication + " does not exist."));
         comment.setCommPub(publication);
         comment.setUserAppComment(user);
         String descriptionFiltree = filtrerMotsInterdits(comment.getDescriptionCommentaire());
@@ -68,16 +78,46 @@ public class ServiceComment implements IServiceComment {
         commentRepository.save(comment);
 
     }
-
-    private String filtrerMotsInterdits(String description) {
+    private String filtrerMotsInterdits(String descriptionCommentaire) {
         // Liste de mots interdits
         List<String> motsInterdits = Arrays.asList("insulte", "haine", "racisme");
 
         for (String motInterdit : motsInterdits) {
-            description = description.replaceAll(motInterdit, "***");
+            descriptionCommentaire = descriptionCommentaire.replaceAll(motInterdit, "***");
         }
 
-        return description;
+        return descriptionCommentaire
+                ;
+    }
+
+    @Override
+    public Comment reportComment(int idComment, long userId) throws Exception {
+        Optional<Comment> commentOptional = commentRepository.findById(idComment);
+        if (!commentOptional.isPresent()) {
+            throw new Exception("Comment with ID " + idComment + " does not exist.");
+        }
+
+        if (!userRepository.existsById(userId)) {
+            throw new Exception("User with ID " + userId + " does not exist.");
+        }
+
+        Comment comment = commentOptional.get();
+        if (comment.getUserAppComment().getUserID() == userId) {
+            throw new Exception("You cannot report your own comment");
+        }
+
+        if (comment.getReportedBy().contains(userId)) {
+            throw new Exception("You have already reported this comment");
+        }
+
+        comment.getReportedBy().add(userId);
+        comment.setSignalCount(comment.getSignalCount() + 1);
+        if (comment.getSignalCount() >= 3) {
+            commentRepository.delete(comment);
+            return null;
+        } else {
+            return commentRepository.save(comment);
+        }
     }
 }
 
@@ -85,9 +125,7 @@ public class ServiceComment implements IServiceComment {
 
 
 
-
-
-    //--------------------------------------------
+//--------------------------------------------
 /*
     @Override
     public Comment addComment(Comment comm) {
